+++ tools/showPawnStructure/tool.js
@@ -300,7 +300,7 @@
       const elems=$(el).find('a[href].mini-game,div.boards>a[href],.study__multiboard a.mini-game,div.mini-game').get();
       if ($(el).is('a[href].mini-game,div.boards>a[href],.study__multiboard a.mini-game,div.mini-game')) elems.push(el[0]);
       for (const el of elems) {
-        fen=fen || $(el).attr('data-state');
+        fen=fen || $(el).attr('data-state') || parent.getPositionFromBoard(el,true);
         if (!fen) {
           parent.global.console.warn('Could not get fen for element',el);
           continue;